revamp parsing of qstat output.  done because some new parsing requires fallbacks; one either has to know the whole list of seen attributes while parsing, or one has to rely on the qstat -f order being fixed.  i found the previous approach less evil.
using the new parsing, one can extract user and group directly from qstat output
when euser and egroup attributes are present.  also, when start_time is present
one can use that.  when any of these are not present, the parser falls back to
the old methods.

a new attribute has been added "startAnchor" which lets the user know which method was
used to compute the start_time.