Use '/usr/bin' for all commands
Replace all occurences of `/sbin/<thing>`, `/bin/<thing>`,
`/usr/sbin/<thing>` with `/usr/bin/<thing>`

The first three are all well on the way to becoming symlinks
to `/usr/bin`.

More over in Fedora 42 rpm macro `%{sbin}` is now `/usr/bin` so this
change is required for Fedora 42 and will be fine on older OSes also.

* https://fedoraproject.org/wiki/Changes/Unify_bin_and_sbin
* https://www.debian.org/releases/bookworm/amd64/release-notes/ch-information.en.html#a-merged-usr-is-now-required
* https://bugzilla.redhat.com/show_bug.cgi?id=2340157